Move note cbox mappings in a separeted function
Since we need to enable the cbox mappings in
different scenarios (e.g. :e *.notes), a default
cbox mapping is moved in an appropriate function, so
we can control the phase that occur when a *BufRead*
or a *BufNewFile* event is received.
This change also includes the ability to enable and
disable the plugin default key mapping and it exposes
a global variable to meet the requirements.
We also changed the default keybindings to add
cbox(es) to something more specific, providing
an example section in the README file.
In order to have a minimum amount of acceptance tests,
this commit adds the cbox checklist that must pass to
allow a PR to get merged.
We improved our automation on PR events adding mergify
with the following rules:
1. Test passed (travis)
2. Approvers >=2

Signed-off-by: fmount <fmount@inventati.org>